requirements.txt
discord.py
python-dotenv
requests
aiohttp
fastapi
uvicorn[standard]
